requirements.txt
mega.py
python-telegram-bot>=20.0